+++ src/rules/no-extraneous-dependencies.js
@@ -11,12 +11,19 @@ function hasKeys(obj = {}) {
   return Object.keys(obj).length > 0
 }
 
+function arrayOrKeys(arrayOrObject) {
+  return Array.isArray(arrayOrObject) ? arrayOrObject : Object.keys(arrayOrObject)
+}
+
 function extractDepFields(pkg) {
   return {
     dependencies: pkg.dependencies || {},
     devDependencies: pkg.devDependencies || {},
     optionalDependencies: pkg.optionalDependencies || {},
     peerDependencies: pkg.peerDependencies || {},
+    // BundledDeps should be in the form of an array, but object notation is also supported by
+    // `npm`, so we convert it to an array if it is an object
+    bundledDependencies: arrayOrKeys(pkg.bundleDependencies || pkg.bundledDependencies || [])
   }
 }
 
@@ -28,6 +35,7 @@ function getDependencies(context, packageDir) {
       devDependencies: {},
       optionalDependencies: {},
       peerDependencies: {},
+      bundledDependencies: [],
     }
 
     if (packageDir && packageDir.length > 0) {
@@ -63,6 +71,7 @@ function getDependencies(context, packageDir) {
       packageContent.devDependencies,
       packageContent.optionalDependencies,
       packageContent.peerDependencies,
+      packageContent.bundledDependencies,
     ].some(hasKeys)) {
       return null
     }
@@ -121,11 +130,13 @@ function reportIfMissing(context, deps, depsOptions, node, name) {
   const isInDevDeps = deps.devDependencies[packageName] !== undefined
   const isInOptDeps = deps.optionalDependencies[packageName] !== undefined
   const isInPeerDeps = deps.peerDependencies[packageName] !== undefined
+  const isInBundledDeps = deps.bundledDependencies.indexOf(packageName) !== -1
 
   if (isInDeps ||
     (depsOptions.allowDevDeps && isInDevDeps) ||
     (depsOptions.allowPeerDeps && isInPeerDeps) ||
-    (depsOptions.allowOptDeps && isInOptDeps)
+    (depsOptions.allowOptDeps && isInOptDeps) ||
+    (depsOptions.allowBundledDeps && isInBundledDeps)
   ) {
     return
   }
@@ -169,6 +180,7 @@ module.exports = {
           'devDependencies': { 'type': ['boolean', 'array'] },
           'optionalDependencies': { 'type': ['boolean', 'array'] },
           'peerDependencies': { 'type': ['boolean', 'array'] },
+          'bundledDependencies': { 'type': ['boolean', 'array'] },
           'packageDir': { 'type': ['string', 'array'] },
         },
         'additionalProperties': false,
@@ -185,6 +197,7 @@ module.exports = {
       allowDevDeps: testConfig(options.devDependencies, filename) !== false,
       allowOptDeps: testConfig(options.optionalDependencies, filename) !== false,
       allowPeerDeps: testConfig(options.peerDependencies, filename) !== false,
+      allowBundledDeps: testConfig(options.bundledDependencies, filename) !== false,
     }
 
     // todo: use module visitor from module-utils core